+++ src/website/assets/scripts/globalConfigs.js
@@ -1,1 +1,1 @@
-window.chiCurrentVersion="6.16.0";
+window.chiCurrentVersion="6.17.0";